style(API): change forlder structure user schema
