+++ packages/api-gateway/src/drivers/http/routes/user/schema/users/getUserSchema.js
@@ -1,0 +1,13 @@
+const getUserSchema = {
+  description: "validate user",
+  tags: ["Users"],
+  params: {
+    type: "object",
+    required: ["userId"],
+    properties: {
+      userId: { type: "string" },
+    },
+  },
+};
+
+module.exports = getUserSchema;
